+++ CHANGELOG.md
@@ -2,6 +2,15 @@
 
 The complete changelog for the Costs to Expect REST API, our changelog follows the format defined at https://keepachangelog.com/en/1.0.0/
 
+## [0.4.0] - 
+### Added
+- Added a share token table to store token for the sharable gaming page links.
+
+### Changed
+- Create a game and adding additional players to a game adds the share token for the player.
+- Redirect to the home page after game creation and adding additional players.
+- Show the link for the public score sheet next to the player name.
+
 ## [0.3.0] - Initial playable release
 
 Initial release of the Yahtzee game scorer which is powered by the Costs to Expect API. If you have an